backend/pyproject.toml
[project]
name = "bart-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "asyncpg",
    "openai",
    "anthropic",
    "pydantic-ai[anthropic]",
    "qdrant-client",
    "python-dotenv",
    "pyngrok",
]
